requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
firebase-admin==6.2.0
deep-translator==1.11.4
Werkzeug==3.0.1
gunicorn==21.2.0

# FastAPI backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Authentication
Pyrebase4==4.7.1
requests-toolbelt==1.0.0
